+++ crates/miden-tx/src/prover/mod.rs
@@ -13,7 +13,11 @@ pub use miden_prover::ProvingOptions;
 use miden_prover::prove;
 
 use super::TransactionProverError;
-use crate::host::{AccountProcedureIndexMap, ScriptMastForestStore};
+use crate::host::{
+    AccountProcedureIndexMap,
+    ScriptMastForestStore,
+    compute_pre_fee_delta_commitment,
+};
 
 mod prover_host;
 pub use prover_host::TransactionProverHost;
@@ -105,21 +109,32 @@ impl LocalTransactionProver {
         )
         .map_err(TransactionProverError::TransactionProgramExecutionFailed)?;
 
-        // extract transaction outputs and process transaction data
-        let (account_delta, output_notes, _tx_progress) = host.into_parts();
+        // Extract transaction outputs and process transaction data.
+        // Note that the account delta already contains the removed transaction fee, so it is the
+        // full delta of the transaction.
+        let (mut post_fee_account_delta, output_notes, _tx_progress) = host.into_parts();
         let tx_outputs =
             TransactionKernel::from_transaction_parts(&stack_outputs, &advice_inputs, output_notes)
                 .map_err(TransactionProverError::TransactionOutputConstructionFailed)?;
 
         // erase private note information (convert private full notes to just headers)
         let output_notes: Vec<_> = tx_outputs.output_notes.iter().map(OutputNote::shrink).collect();
-        let account_delta_commitment = account_delta.to_commitment();
+
+        // Because the fee asset is removed from the vault after the commitment is computed in the
+        // kernel, we have to *add* it to the delta before computing the commitment against which
+        // the transaction is proven.
+        // Note that the fee asset is a transaction output and so is part of the proof. The delta
+        // without the added fee asset can still be validated by repeating this process.
+        // Compute the delta commitment of the delta before the fee was removed.
+        let pre_fee_delta_commitment =
+            compute_pre_fee_delta_commitment(&mut post_fee_account_delta, tx_outputs.fee)
+                .map_err(TransactionProverError::ComputePreFeeDelta)?;
 
         let builder = ProvenTransactionBuilder::new(
             account.id(),
             account.init_commitment(),
             tx_outputs.account.commitment(),
-            account_delta_commitment,
+            pre_fee_delta_commitment,
             ref_block_num,
             ref_block_commitment,
             tx_outputs.fee,
@@ -135,12 +150,12 @@ impl LocalTransactionProver {
                 let account_update_details = if account.is_new() {
                     let mut account = account.clone();
                     account
-                        .apply_delta(&account_delta)
+                        .apply_delta(&post_fee_account_delta)
                         .map_err(TransactionProverError::AccountDeltaApplyFailed)?;
 
                     AccountUpdateDetails::New(account)
                 } else {
-                    AccountUpdateDetails::Delta(account_delta)
+                    AccountUpdateDetails::Delta(post_fee_account_delta)
                 };
 
                 builder.account_update_details(account_update_details)